feat: split ssg test
