--- tests/integration/ssg/tsconfig.json
@@ -1,14 +1,0 @@
-{
-  "extends": "@modern-js/tsconfig/base",
-  "compilerOptions": {
-    "declaration": false,
-    "jsx": "react-jsx",
-    "baseUrl": "./",
-    "paths": {
-      "@/*": ["./src/*"],
-      "@shared/*": ["./shared/*"]
-    },
-    "types": ["node", "jest"]
-  },
-  "include": ["tests"]
-}
